Modificacion estructura, cambio tipo variables, completo MYSQLFacturaDAO y agrego docker

+++ integrador-1/src/main/java/integrador_1/entity/Factura.java
@@ -8,6 +8,6 @@ import lombok.*;
 @Setter
 @ToString
 public class Factura {
-    private Long idFactura;
-    private Long idCliente;    
+    private int idFactura;
+    private int idCliente;
 }
